// client/src/generated-client.ts
export class AuthorDto implements IAuthorDto {
    id?: string;
    name?: string;
    createdAt?: Date | undefined;
    booksIds?: string[];

    constructor(data?: IAuthorDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.createdAt = _data["createdAt"] ? new Date(_data["createdAt"].toString()) : undefined as any;
            if (Array.isArray(_data["booksIds"])) {
                this.booksIds = [] as any;
                for (let item of _data["booksIds"])
                    this.booksIds!.push(item);
            }
        }
    }

    static fromJS(data: any): AuthorDto {
        data = typeof data === 'object' ? data : {};
        let result = new AuthorDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["createdAt"] = this.createdAt ? this.createdAt.toISOString() : undefined as any;
        if (Array.isArray(this.booksIds)) {
            data["booksIds"] = [];
            for (let item of this.booksIds)
                data["booksIds"].push(item);
        }
        return data;
    }
}

export interface IAuthorDto {
    id?: string;
    name?: string;
    createdAt?: Date | undefined;
    booksIds?: string[];
}
// ...
export class BookDto implements IBookDto {
    id?: string;
    title?: string;
    pages?: number;
    createdat?: Date | undefined;
    genre?: GenreDto | undefined;
    authorsIds?: string[];
    authors?: AuthorDto[];

    constructor(data?: IBookDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.title = _data["title"];
            this.pages = _data["pages"];
            this.createdat = _data["createdat"] ? new Date(_data["createdat"].toString()) : undefined as any;
            this.genre = _data["genre"] ? GenreDto.fromJS(_data["genre"]) : undefined as any;
            if (Array.isArray(_data["authorsIds"])) {
                this.authorsIds = [] as any;
                for (let item of _data["authorsIds"])
                    this.authorsIds!.push(item);
            }
            if (Array.isArray(_data["authors"])) {
                this.authors = [] as any;
                for (let item of _data["authors"])
                    this.authors!.push(AuthorDto.fromJS(item));
            }
        }
    }

    static fromJS(data: any): BookDto {
        data = typeof data === 'object' ? data : {};
        let result = new BookDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["title"] = this.title;
        data["pages"] = this.pages;
        data["createdat"] = this.createdat ? this.createdat.toISOString() : undefined as any;
        data["genre"] = this.genre ? this.genre.toJSON() : undefined as any;
        if (Array.isArray(this.authorsIds)) {
            data["authorsIds"] = [];
            for (let item of this.authorsIds)
                data["authorsIds"].push(item);
        }
        if (Array.isArray(this.authors)) {
            data["authors"] = [];
            for (let item of this.authors)
                data["authors"].push(item ? item.toJSON() : undefined as any);
        }
        return data;
    }
}

export interface IBookDto {
    id?: string;
    title?: string;
    pages?: number;
    createdat?: Date | undefined;
    genre?: GenreDto | undefined;
    authorsIds?: string[];
    authors?: AuthorDto[];
}

export class GenreDto implements IGenreDto {
    id?: string;
    name?: string;
    createdAt?: Date | undefined;
    books?: string[];

    constructor(data?: IGenreDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.createdAt = _data["createdAt"] ? new Date(_data["createdAt"].toString()) : undefined as any;
            if (Array.isArray(_data["books"])) {
                this.books = [] as any;
                for (let item of _data["books"])
                    this.books!.push(item);
            }
        }
    }

    static fromJS(data: any): GenreDto {
        data = typeof data === 'object' ? data : {};
        let result = new GenreDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["createdAt"] = this.createdAt ? this.createdAt.toISOString() : undefined as any;
        if (Array.isArray(this.books)) {
            data["books"] = [];
            for (let item of this.books)
                data["books"].push(item);
        }
        return data;
    }
}

export interface IGenreDto {
    id?: string;
    name?: string;
    createdAt?: Date | undefined;
    books?: string[];
}

export class CreateBookRequestDto implements ICreateBookRequestDto {
    title!: string;
    pages!: number;
    authorsIds?: string[];
    genreId?: string | undefined;

    constructor(data?: ICreateBookRequestDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.title = _data["title"];
            this.pages = _data["pages"];
            if (Array.isArray(_data["authorsIds"])) {
                this.authorsIds = [] as any;
                for (let item of _data["authorsIds"])
                    this.authorsIds!.push(item);
            }
            this.genreId = _data["genreId"];
        }
    }

    static fromJS(data: any): CreateBookRequestDto {
        data = typeof data === 'object' ? data : {};
        let result = new CreateBookRequestDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["title"] = this.title;
        data["pages"] = this.pages;
        if (Array.isArray(this.authorsIds)) {
            data["authorsIds"] = [];
            for (let item of this.authorsIds)
                data["authorsIds"].push(item);
        }
        data["genreId"] = this.genreId;
        return data;
    }
}

export interface ICreateBookRequestDto {
    title: string;
    pages: number;
    authorsIds?: string[];
    genreId?: string | undefined;
}

export class UpdateBookRequestDto implements IUpdateBookRequestDto {
    bookIdForLookupReference!: string;
    newTitle!: string;
    newPageCount!: number;
    authorsIds!: string[];
    genreId?: string | undefined;

    constructor(data?: IUpdateBookRequestDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
        if (!data) {
            this.authorsIds = [];
        }
    }

    init(_data?: any) {
        if (_data) {
            this.bookIdForLookupReference = _data["bookIdForLookupReference"];
            this.newTitle = _data["newTitle"];
            this.newPageCount = _data["newPageCount"];
            if (Array.isArray(_data["authorsIds"])) {
                this.authorsIds = [] as any;
                for (let item of _data["authorsIds"])
                    this.authorsIds!.push(item);
            }
            this.genreId = _data["genreId"];
        }
    }

    static fromJS(data: any): UpdateBookRequestDto {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateBookRequestDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["bookIdForLookupReference"] = this.bookIdForLookupReference;
        data["newTitle"] = this.newTitle;
        data["newPageCount"] = this.newPageCount;
        if (Array.isArray(this.authorsIds)) {
            data["authorsIds"] = [];
            for (let item of this.authorsIds)
                data["authorsIds"].push(item);
        }
        data["genreId"] = this.genreId;
        return data;
    }
}

export interface IUpdateBookRequestDto {
    bookIdForLookupReference: string;
    newTitle: string;
    newPageCount: number;
    authorsIds: string[];
    genreId?: string | undefined;
}